Move pytest fixtures into common conftest file
