+++ tests/test_chips.py
@@ -3,16 +3,6 @@ import pytest
 from blackjack.chips import Chips
 
 
-@pytest.fixture
-def chips():
-    """Create chips class for testing
-
-    Returns:
-        _type_: an instance of a class of chips
-    """
-    return Chips()
-
-
 def test_win_bet(chips: Chips):
     """Test the chips total increases if bet is won
